generate: provide a schema override method for matching
While pretty close, not all fields are 1:1 between the API and the schema. This
leads to scenarios where the attribute name is either: 1) never set in the API
response or 2) conditionally set by the CRUD operations due to conflicts or
self contained logic. See `cloudflare_record`[1] for an example of this.

[1]: https://github.com/cloudflare/terraform-provider-cloudflare/blob/b55e604df70794bbf6dbfc5e425f8a2e30b7b6a1/cloudflare/resource_cloudflare_record.go#L306-L309