cloudflare_argo: add initial support
Introduces support for `cloudflare_argo` to be exported and imported. This
resource is a little different from most (but definitely not unique) as it
makes multiple API requests in order to populate the resource. This makes it
really hard to test in the current setup despite being a very simple resource.