generate: support remapping rulesets overrides `enabled` => `status`
cloudflare/terraform-provider-cloudflare#1689 introduced a breaking change where
the `status` field of an override is now `enabled` in order to support boolean
values that can be true, false and uninitialised in the API (see PR for full
details).

This ensures that the mapping of these values is performed on the generated
resources too.