util: prevent --account and --zone being used together
These flags are mutually exclusive and are used to determine which level
(zone or account) resources should be exported. The previous behaviour
would discard zone in the event of a collision.

Last time we tried this in #356, we still had the dependency on the
global `account_id`. That has now been removed (since cloudflare-go
0.57.0[1]) and now requires explicitly identifiers for all Workers
endpoints. Worker Routes still use the zone levels APIs while everything
else is account level based.

[1]: https://github.com/cloudflare/cloudflare-go/blob/master/CHANGELOG.md#0570-december-22nd-2022