generate/cloudflare_tunnel: manually provide pagination information
Due to the API service not responding with a correct `result_info`
section of the payload, automatic pagination is not available. To work
around this, we are explicitly only fetching the first 1000 active
Cloudflare Tunnels.

This is in the team's backlog internally and can be removed once it is
addressed.