generate/load_balancer: correctly remap `country_pools`, `region_pools` and `pop_pools`
Updates the `cloudflare_load_balancer` generation to remap the
`country_pools`, `region_pools` and `pop_pools` to the nested block
instead of the object/map field.

This was a bit of tricky one to track down because the original response
from the API _has_ the correct schema attributes but the kicker is that
they are not a supported type (for good reasons!).