generate: remove additional handling of numeric values
In previous versions of our schema converter, we needed to handle the scenario
where a key inside of a map would contain a numeric value.

Note: This is only required for `TypeMap`s because it is the wild west and you
lose most of the validation with the Terraform internals when using maps.

```
resource "..." "..." {
  foo = {
    bar1 = "baz"
  }
}
```

While numeric characters in attribute keys are permitted in HCL, they must be
quoted. Inadvertently, it looks like this was also being applied to the
attribute values. While a good measure, this isn't as important as we have
better measures to detect the value type and infer it appropriately.

This commit removes the additional handling and instead relies on the explicit
`.(type)` check instead which covers the key _and_ value.