refactor(internal): swap to resource and endpoint mapping
Previously, we relied on manually mapping the supported resources. This
was for a number of reasons:

1. Not all resources supported a `List` operation.
2. Resources were not 1:1 to the API endpoint in Terraform.
3. Resources did not have full attribute coverage and not having a way
   to flag incomplete resources could result in outages if they were
   applied.

With the release of the Cloudflare Terraform Provider v5 and the SDKs,
we're now in a position to make the internals more dynamic as they are
all powered by the OpenAPI schemas and eliminates the previously
encountered constraints.

Under the covers, we have swapped to a mapping file that is generated by
an out of band process while the HTTP client and data handling is now
using the v4 Go library.