chore: handle non-standard registries more automatically
When everyone was upset with Hashicorp and creating their own
registries, it was the wild west and we didn't really have a great way
to universally detect the registry in use without it being explicitly
provided.

Fast forward to today, everyone* now ships this as part of their state
implementation (like Hashicorp always did) so it's easier to determine
the registry once it is initialised.

_* Well, anyone that has a > 0.01% footprint for usage._